notebook-vision/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
easyocr==1.7.2
Pillow==11.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
